Gets rid of classmethods that aren't necessary.
